Fixed upload images
Fixed:
- Upload images
- CSS styles with images
- Reset of backend images (exception if contains cute, meme or chibi)
- Fixed some bugs with JS console in "Projects"

+++ server-app-java/src/main/java/com/iman/service/users/UserService.java
@@ -156,7 +156,11 @@ public class UserService {
 		
 		// delete from path previous image
 		if(StringUtils.isNotBlank(user.getImageUid())) {
-			deleteImage(path, user);
+			try {
+				deleteImage(path, user);
+			} catch (Exception e) {
+				// TODO: handle exception
+			}
 		}
 		
 		// create new image & update
@@ -179,7 +183,11 @@ public class UserService {
 	@Transactional
 	public void deleteImage(String path, User user) {
 		try {
-			FileUtils.deleteFromPath(path, user.getImageUid());
+			if(!user.getImageUid().contains("cute") && !user.getImageUid().contains("meme") && !user.getImageUid().contains("chibi")) {
+				FileUtils.deleteFromPath(path, user.getImageUid());
+			}
+			user.setImageUid(null);
+			userRepository.save(user);
 		} catch (IOException e) {
 			throw new RuntimeException();
 		}
